create new package scan for pysm operator
closes #200
