+++ src/python/tod/__init__.py
@@ -20,7 +20,7 @@ from .sim_interval import regular_intervals
 
 from .sim_det_noise import OpSimNoise
 
-from .sim_det_map import OpSimGradient, OpSimScan, OpSimPySM
+from .sim_det_map import OpSimGradient, OpSimScan
 
 from .sim_det_dipole import OpSimDipole
